Magisk delta and kitsune logo latest 26.3 update

+++ native/src/init/rootdir.cpp
@@ -79,13 +79,13 @@ on property:init.svc.zygote=stopped
     clone_attr(src, dest);
 }
 
-static void load_overlay_rc(const char *overlay) {
+static void load_overlay_rc(const char *overlay, bool should_unlink = true) {
     auto dir = open_dir(overlay);
     if (!dir) return;
 
     int dfd = dirfd(dir.get());
     // Do not allow overwrite init.rc
-    unlinkat(dfd, "init.rc", 0);
+    if (should_unlink) unlinkat(dfd, "init.rc", 0);
 
     // '/' + name + '\0'
     char buf[NAME_MAX + 2];
@@ -102,7 +102,7 @@ static void load_overlay_rc(const char *overlay) {
             int rc = xopenat(dfd, entry->d_name, O_RDONLY | O_CLOEXEC);
             rc_list.push_back(full_read(rc));
             close(rc);
-            unlinkat(dfd, entry->d_name, 0);
+            if (should_unlink) unlinkat(dfd, entry->d_name, 0);
         }
     }
 }
@@ -158,6 +158,14 @@ static void magic_mount(const string &sdir, const string &ddir = "") {
     }
 }
 
+static void patch_socket_name(const char *path) {
+    static char rstr[16] = { 0 };
+    if (rstr[0] == '\0')
+        random_strc(rstr, sizeof(rstr));
+    auto bin = mmap_data(path, true);
+    bin.patch(MAIN_SOCKET, rstr);
+}
+
 static void extract_files(bool sbin) {
     const char *m32 = sbin ? "/sbin/magisk32.xz" : "magisk32.xz";
     const char *m64 = sbin ? "/sbin/magisk64.xz" : "magisk64.xz";
@@ -167,17 +175,17 @@ static void extract_files(bool sbin) {
         mmap_data magisk(m32);
         unlink(m32);
         int fd = xopen("magisk32", O_WRONLY | O_CREAT, 0755);
-        fd_channel ch(fd);
-        unxz(ch, magisk);
+        unxz(fd, magisk.buf(), magisk.sz());
         close(fd);
+        patch_socket_name("magisk32");
     }
     if (access(m64, F_OK) == 0) {
         mmap_data magisk(m64);
         unlink(m64);
         int fd = xopen("magisk64", O_WRONLY | O_CREAT, 0755);
-        fd_channel ch(fd);
-        unxz(ch, magisk);
+        unxz(fd, magisk.buf(), magisk.sz());
         close(fd);
+        patch_socket_name("magisk64");
         xsymlink("./magisk64", "magisk");
     } else {
         xsymlink("./magisk32", "magisk");
@@ -186,8 +194,7 @@ static void extract_files(bool sbin) {
         mmap_data stub(stub_xz);
         unlink(stub_xz);
         int fd = xopen("stub.apk", O_WRONLY | O_CREAT, 0);
-        fd_channel ch(fd);
-        unxz(ch, stub);
+        unxz(fd, stub.buf(), stub.sz());
         close(fd);
     }
 }
@@ -254,7 +261,7 @@ void MagiskInit::patch_ro_root() {
     }
 #endif
 
-    load_overlay_rc(ROOTOVL);
+    load_overlay_rc(INTLROOT "/early-mount.d/initrc.d", false);
     if (access(ROOTOVL "/sbin", F_OK) == 0) {
         // Move files in overlay.d/sbin into tmp_dir
         mv_path(ROOTOVL "/sbin", ".");
@@ -313,10 +320,6 @@ void MagiskInit::patch_rw_root() {
     rm_rf("/data/overlay.d");
     rm_rf("/.backup");
 
-    // Patch init.rc
-    patch_init_rc("/init.rc", "/init.p.rc", "/sbin");
-    rename("/init.p.rc", "/init.rc");
-
     bool treble;
     {
         auto init = mmap_data("/init");
@@ -327,6 +330,14 @@ void MagiskInit::patch_rw_root() {
     xmount("tmpfs", PRE_TMPSRC, "tmpfs", 0, "mode=755");
     xmkdir(PRE_TMPDIR, 0);
     setup_tmp(PRE_TMPDIR);
+
+    // Handle custom rc script
+    load_overlay_rc(PRE_TMPDIR "/" MIRRDIR "/early-mount/initrc.d", false);
+
+    // Patch init.rc
+    patch_init_rc("/init.rc", "/init.p.rc", "/sbin");
+    rename("/init.p.rc", "/init.rc");
+
     chdir(PRE_TMPDIR);
 
     // Extract magisk
